investment_analyzer/requirements.txt
futu-api>=6.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
streamlit>=1.25.0
openpyxl>=3.1.0
python-dotenv>=1.0.0